backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
groq==0.13.1
python-dotenv==1.0.1
pydantic==2.9.2
aiosqlite==0.20.0
httpx==0.27.2
python-multipart==0.0.12
dnspython==2.7.0
email-validator==2.2.0
